LayoutTests/imported/w3c/web-platform-tests/html/anonymous-iframe/session-storage.tentative.https.window-expected.txt


PASS Setup
FAIL same_origin anonymous iframe can't access the sessionStorage assert_equals: expected "" but got "same_origin"
PASS cross_origin anonymous iframe can't access the sessionStorage
